+++ crates/orderbook/src/dto/mod.rs
@@ -1,0 +1,7 @@
+pub mod auction;
+pub mod order;
+
+pub use {
+    auction::{Auction, AuctionId, AuctionWithId},
+    order::Order,
+};
