+++ crates/orderbook/src/dto/order.rs
@@ -1,0 +1,44 @@
+use {
+    model::{
+        app_data::AppDataHash,
+        interaction::InteractionData,
+        order::{BuyTokenDestination, OrderClass, OrderKind, OrderUid, SellTokenSource},
+        signature::Signature,
+    },
+    number::serialization::HexOrDecimalU256,
+    primitive_types::{H160, U256},
+    serde::{Deserialize, Serialize},
+    serde_with::serde_as,
+};
+
+#[serde_as]
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Order {
+    pub uid: OrderUid,
+    pub sell_token: H160,
+    pub buy_token: H160,
+    #[serde_as(as = "HexOrDecimalU256")]
+    pub sell_amount: U256,
+    #[serde_as(as = "HexOrDecimalU256")]
+    pub buy_amount: U256,
+    #[serde_as(as = "HexOrDecimalU256")]
+    pub solver_fee: U256,
+    #[serde_as(as = "HexOrDecimalU256")]
+    pub user_fee: U256,
+    pub valid_to: u32,
+    pub kind: OrderKind,
+    pub receiver: Option<H160>,
+    pub owner: H160,
+    pub partially_fillable: bool,
+    #[serde_as(as = "HexOrDecimalU256")]
+    pub executed: U256,
+    pub pre_interactions: Vec<InteractionData>,
+    pub post_interactions: Vec<InteractionData>,
+    pub sell_token_balance: SellTokenSource,
+    pub buy_token_balance: BuyTokenDestination,
+    pub class: OrderClass,
+    pub app_data: AppDataHash,
+    #[serde(flatten)]
+    pub signature: Signature,
+}
