+++ crates/orderbook/src/lib.rs
@@ -2,6 +2,7 @@ pub mod api;
 pub mod app_data;
 pub mod arguments;
 pub mod database;
+pub mod dto;
 mod ipfs;
 mod ipfs_app_data;
 pub mod orderbook;
